+++ website/static/playground/worker.js
@@ -34,6 +34,7 @@ async function instantiate() {
         host_dom_get_value: () => 0,
         host_dom_set_value: () => {},
         host_dom_on: () => {},
+        host_dom_off: () => 0,
         host_dom_focus: () => {},
         host_dom_set_class: () => {},
         host_dom_fetch: () => {},